WIP noninteractive CLI
