+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "vger"
-version = "0.1.6"
+version = "0.1.7"
 description = "An execution framework for Jupyter environments."
 authors = ["Joseph Lucas <joe@joetl.com>"]
 license = "GPL-3.0-only"
@@ -20,6 +20,7 @@ rich = "^13.7"
 urllib3 = "^2.1"
 websockets = "^12.0"
 requests = "^2.30"
+fire = "^0.5.0"
